requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
psycopg2-binary==2.9.9
pymongo==4.6.1
pyyaml==6.0.1
pandas==2.2.0
streamlit==1.29.0
storm
streamparse
streamlit 
plotly